Move each command handler's unit tests into their own files
The test layout now mirrors the source layout: each handler under
src/interpreter/commands/ has a matching test file under
src/__tests__/interpreter/commands/. The interpreter test file keeps
what belongs to the interpreter itself: execute(), the registry
identity tests, and the whole-pipeline malformed-coordinates suite.
Pure move; test names and bodies are unchanged.
